Use zip -O option so that ja_JP.SJIS locale is not required
ja_JP.SJIS locale is not supported in Red Hat officially.
And another solution is to use 'zip -O' option.
'zip -O' is a Red Hat internal patch to work on UTF-8 locale.

This way can be used in Red Hat 6 or Fedora only and cannot apply to
other arhive types because the -O option is ZIP specific.